# Cache expensive data in `pl-code` element