--- a/apps/prairielearn/elements/pl-code/pl-code.py
+++ b/apps/prairielearn/elements/pl-code/pl-code.py
@@ -1,4 +1,5 @@
 import os
+from functools import cache
 from html import escape, unescape
 from typing import Any, Generator, Iterable, Iterator, Optional
 
@@ -9,6 +10,7 @@ import pygments
 import pygments.formatters
 import pygments.lexer
 import pygments.lexers
+import pygments.style
 import pygments.util
 from code_utils import parse_highlight_lines
 from pygments.styles import STYLE_MAP, get_style_by_name
@@ -47,6 +49,10 @@ ANSI_COLORS = {
 }
 
 
+# Computing this is relatively expensive, so we'll do it once here and reuse it.
+ansi_color_tokens = color_tokens(ANSI_COLORS, ANSI_COLORS)
+
+
 class NoHighlightingLexer(pygments.lexer.Lexer):
     """
     Dummy lexer for when syntax highlighting is not wanted, but we still
@@ -68,6 +74,12 @@ class HighlightingHtmlFormatter(pygments.formatters.HtmlFormatter):
     Subclass of the default HTML formatter to provide more flexibility
     with highlighted lines.
     """
+
+    def set_highlighted_lines(self, hl_lines: Optional[list[int]]) -> None:
+        self.hl_lines.clear()
+        if hl_lines:
+            for line in hl_lines:
+                self.hl_lines.add(line)
 
     def _highlight_lines(
         self, tokensource: Iterable[tuple[int, str]]
@@ -93,10 +105,16 @@ class HighlightingHtmlFormatter(pygments.formatters.HtmlFormatter):
         return ""
 
 
+# Selecting and instantiating a lexer is actually reasonably expensive
+# (on the order of 5-10ms). We cache a resolution from a language name to a
+# lexers to avoid this cost if `<pl-code>` is used with the same language
+# multiple times in the same question.
+@cache
 def get_lexer_by_name(name: str) -> Optional[pygments.lexer.Lexer]:
     """
     Tries to find a lexer by both its proper name and any aliases it has.
     """
+
     # Search by proper class/language names
     # This returns None if not found, and a class if found.
     lexer_class = pygments.lexers.find_lexer_class(name)
@@ -110,6 +128,31 @@ def get_lexer_by_name(name: str) -> Optional[pygments.lexer.Lexer]:
             return pygments.lexers.get_lexer_by_name(name)
         except pygments.util.ClassNotFound:
             return None
+
+
+@cache
+def get_formatter(
+    BaseStyle: type[pygments.style.Style], highlight_lines_color: Optional[str]
+):
+    class CustomStyleWithAnsiColors(BaseStyle):
+        styles = dict(BaseStyle.styles)
+        styles.update(ansi_color_tokens)
+
+        highlight_color = (
+            highlight_lines_color or BaseStyle.highlight_color or "#b3d7ff"
+        )
+
+    formatter_opts = {
+        "style": CustomStyleWithAnsiColors,
+        "nobackground": True,
+        "noclasses": True,
+        # We'll unconditionally render the line numbers, but we'll hide them if
+        # they aren't specifically enabled. This means we only have to deal with
+        # one markup "shape" in our CSS, not two.
+        "linenos": "table",
+    }
+
+    return HighlightingHtmlFormatter(**formatter_opts)
 
 
 def prepare(element_html: str, data: pl.QuestionData) -> None:
@@ -224,28 +267,10 @@ def render(element_html: str, data: pl.QuestionData) -> str:
     background_color = pygments_style.background_color or "transparent"
     line_number_color = pygments_style.line_number_color
 
-    class CustomStyleWithAnsiColors(pygments_style):
-        styles = dict(pygments_style.styles)
-        styles.update(color_tokens(ANSI_COLORS, ANSI_COLORS))
-
-        highlight_color = (
-            pygments_style.highlight_color or highlight_lines_color or "#b3d7ff"
-        )
-
-    formatter_opts = {
-        "style": CustomStyleWithAnsiColors,
-        "nobackground": True,
-        "noclasses": True,
-        # We'll unconditionally render the line numbers, but we'll hide them if
-        # they aren't specifically enabled. This means we only have to deal with
-        # one markup "shape" in our CSS, not two.
-        "linenos": "table",
-    }
-
-    if highlight_lines is not None:
-        formatter_opts["hl_lines"] = parse_highlight_lines(highlight_lines)
-
-    formatter = HighlightingHtmlFormatter(**formatter_opts)
+    formatter = get_formatter(pygments_style, highlight_lines_color)
+    formatter.set_highlighted_lines(
+        parse_highlight_lines(highlight_lines) if highlight_lines else None
+    )
 
     code = pygments.highlight(unescape(code), lexer, formatter)