--- a/exampleCourse/questions/element/code/question.html
+++ b/exampleCourse/questions/element/code/question.html
@@ -28,7 +28,7 @@ def square(x):
 <pl-question-panel>
 <markdown>
 Another useful feature of `pl-code` is the ability to highlight specific lines. This is done by using
-`highlight-lines="low-high"` attribute. Moreover, one can disable selecting the code region by
+`highlight-lines="..."` attribute. Moreover, one can disable selecting the code region by
 adding `prevent-select="true"`.
 </markdown>
 <pl-code language="python" highlight-lines="4-5" prevent-select="true">