This PR builds on the changes from #10759 to improve the performance of questions that use a lot of `pl-code` elements with the same language/style.

The following numbers are using the `matlab/practice assessment questions/sas_checkbox/image_manipulation/crop_and_highlight` question from UMich's ENGR 101 course. Testing was done with `process-questions-in-worker` feature enabled.

Master:

```
freeform.generate: 15.00ms
freeform.prepare: 505.75ms
freeform.render: 874.15ms
freeform.generate: 13.43ms
freeform.prepare: 480.76ms
freeform.render: 898.65ms
freeform.generate: 10.20ms
freeform.prepare: 476.02ms
freeform.render: 879.08ms
freeform.generate: 14.67ms
freeform.prepare: 479.53ms
freeform.render: 877.84ms
freeform.generate: 15.20ms
freeform.prepare: 475.83ms
freeform.render: 874.03ms
```

On this branch:

```
freeform.generate: 11.11ms
freeform.prepare: 286.16ms
freeform.render: 291.82ms
freeform.generate: 10.55ms
freeform.prepare: 263.29ms
freeform.render: 284.92ms
freeform.generate: 15.15ms
freeform.prepare: 264.59ms
freeform.render: 287.95ms
freeform.generate: 16.61ms
freeform.prepare: 274.62ms
freeform.render: 283.91ms
freeform.generate: 15.70ms
freeform.prepare: 272.00ms
freeform.render: 283.25ms
```

Summary (blame ChatGPT for any wrong numbers):

| operation | `master` average | branch average | diff |
|--------|--------|--------|--------|
| `generate` | 13.70ms | 13.82ms | -1% |
| `prepare` | 483.57ms | 272.13ms | 44% |
| `render` | 880.75ms | 286.37ms | 67% |

This is a very very good improvement: overall, we're now almost 60% faster on the initial generate/prepare/render for this question.

This also yields a smaller gain for the `element/code` question in `exampleCourse`, with the full generate/prepare/render running 70ms faster (roughly a 12% decrease). This is a smaller win because it's using a wider variety of languages/styles so caching doesn't matter as much.